Enable Errors support for any multi-error
Starting Go 1.20, any multi-error should conform to the standard
unwrap method: Unwrap() []error.

This changes multierr.Errors() method to support any error that
complies to that interface.

Fix #70.
